+++ binac/mpi4py_cupy/example5.py
@@ -5,8 +5,8 @@ comm = MPI.COMM_WORLD
 size = comm.Get_size()
 rank = comm.Get_rank()
 
-sendbuf = cp.arange(10, dtype=cp.float64)
-recvbuf = cp.empty_like(sendbuf)
+sendbuf = cp.arange(10, dtype=float)
+recvbuf = cp.empty_like(sendbuf, dtype=float)
 assert hasattr(sendbuf, '__cuda_array_interface__')
 assert hasattr(recvbuf, '__cuda_array_interface__')
 cp.cuda.get_current_stream().synchronize()
